fix: moving use-cases folder
