+++ backend/src/presentation/controllers/courseController.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
 
 import { CreateCourseDTO } from "../../application/dtos/course/createCourse.dto";
-import { CreateCourse } from "../../application/use-cases/course/createCourse";
+import { CreateCourse } from "../../application/useCases/course/createCourse";
 import { UpdateCourseDTO } from "../../application/dtos/course/updateCourse.dto";
-import { UpdateCourse } from "../../application/use-cases/course/updateCourse";
+import { UpdateCourse } from "../../application/useCases/course/updateCourse";
 import { ListCourseDTO } from "../../application/dtos/course/listCourse.dto";
-import { ListCourse } from "../../application/use-cases/course/listCourse";
+import { ListCourse } from "../../application/useCases/course/listCourse";
 export class CourseController {
   constructor(
     private createCourse: CreateCourse,
